backend/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
sqlalchemy==2.0.36
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pyjwt==2.10.1
httpx==0.28.0
psycopg2-binary==2.9.10